run long running git log command in parallel
The git command to get the first commit that introduced the sysfs
attribute looks like this:

git log -L $(line_num),$(line_num):$(filename) --reverse

For a very large codebase like the kernel, it is excruciatingly _slow_
but almost always gives the correct result. (Need to re-check whether it
fails when lines are moved in the same file.)

Used python's multiprocessing module to run the git tasks in parallel.
This is bound to take up 100% of the CPU.

Closes #3

Signed-off-by: Aishwarya Pant <aishpant@gmail.com>